+++ src/pages/AuctionOfferAcceptedConfirmationPage.tsx
@@ -25,7 +25,6 @@ import { driverHighResImage } from '../constants/images';
 import useIsMounted from '../hooks/use-is-mounted';
 import { AuthLoginContext } from '../providers/AuthLoginProvider';
 import {AcceptBidAction } from '../store/actions/auctions';
-import AuctionsMyAuctionsPage from './AuctionsMyAuctionsPage';
 import { auctionSelector } from '../store/selectors/auctions';
 import { CollectiblesOnAuctionLoadAction } from '../store/actions/collections';
 import { transactionsSelector } from '../store/selectors/transaction';
@@ -96,12 +95,6 @@ const AuctionOfferAcceptedConfirmationPage: FC = () => {
   const transactions = useSelector(transactionsSelector, shallowEqual);
   const tx = useMemo(() => transactions[txUuid], [transactions]);
 
-  const [navToMyAuctions, setNavToMyAuctions] = useState(false);
-
-  const navigateMyAuctions = () => {    
-    setNavToMyAuctions(true);   
-  }     
-
   const isLoading = useCallback(() => {
     return acceptOfferRequest?.isLoading || tx?.isLoading;
   }, [acceptOfferRequest, tx]);
@@ -127,11 +120,6 @@ const AuctionOfferAcceptedConfirmationPage: FC = () => {
   }, [isMounted, dispatch, bidId, auctionId, publicKey, passphrase]);  
 
   return (
-    <>
-    {navToMyAuctions && (
-      <AuctionsMyAuctionsPage />
-    )} 
-    {!navToMyAuctions && ( 
     <IonPage>
       <Header 
         title="Accept Bid Offer"
@@ -194,7 +182,6 @@ const AuctionOfferAcceptedConfirmationPage: FC = () => {
               expand="block"
               onClick={() => {
                   dispatch(CollectiblesOnAuctionLoadAction(publicKey!, true, false, true, undefined));
-                  //navigateMyAuctions();
                   history.push('/market/myauctions');
                 }
               } 
@@ -204,9 +191,7 @@ const AuctionOfferAcceptedConfirmationPage: FC = () => {
           </IonToolbar>
         </Footer>
       )}
-    </IonPage>
-    )}  
-    </>    
+    </IonPage>  
   );
 };
